requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
mysqlclient>=2.2.0
python-dotenv>=1.0.1
requests>=2.32.0